(refactor) Add controller to routes
