+++ server/index.js
@@ -1,18 +1,16 @@
 var express = require('express');
-var roomRouter = require('./routers/room.js');
-var roomyRouter = require('./routers/roomy.js');
-
+var app = express();
+var router = require('./routes.js');
+var db = require('./db');
 
 var port = 8000;
-var app = express();
-// connect to database here
-
 
 // use middleware here
-app.use(express.static(__dirname + '/../src'));
-app.use('/api/room', roomRouter); // adjust router name and endpoint when necessary
-app.use('/api/roomy', roomyRouter); // adjust router name and endpoint when necessary
 
+//possibly change endpoint 
+app.use('/api/roomy', router); 
+
+app.use(express.static(__dirname + '/../src'));
 
 app.listen(port, function() {
   console.log('Listening on port: ' + port);
