+++ server/routes.js
@@ -1,0 +1,9 @@
+var controller = require('./controller');
+var router = require('express').Router();
+
+router.get('/users', controller.users.get);
+
+router.post('/users', controller.users.post);
+
+
+module.exports = router;
