--- server/routes/room.js
@@ -1,11 +1,0 @@
-var express = require('express');
-var roomController = require('../../db/controllers/room.js');
-
-var router = express.Router();
-
-router.route('/')
-  .get(function(req, res) {
-    // use roomController methods here
-  });
-
-module.exports = router;
